requirements.txt
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
lxml==4.5.2
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
queuelib==1.5.0
Scrapy==2.2.0
scrapy-djangoitem==1.1.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
zope.interface==5.1.0
